Move view refresh to transaction views
